Remove old PostMessage class, rename AbstractPostMessage to PostMessage, add default page_content method to PostMessage
